fix: package json scripts and folder structure
