+++ server/package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "main": "./index.ts",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development concurrently \"npx tsc --watch\" \" nodemon -q dist/index.js\"",
+    "build": "npx tsc",
+    "dev": "nodemon ./src/index.ts",
     "test": "jest",
     "db:push:dev": "cross-env NODE_ENV=development node dist/db/dbpush.js",
     "db:push:test": "cross-env NODE_ENV=testing node dist/db/dbpush.js"
@@ -31,6 +32,7 @@
   "dependencies": {
     "@babel/core": "^7.24.7",
     "@prisma/client": "^5.15.0",
+    "@types/node": "^20.14.6",
     "bcryptjs": "^2.4.3",
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.4.5",
