+++ server/src/controllers/authControllers/logOutUser.controller.ts
