+++ server/src/controllers/authControllers/signUpUser.controller.ts
