+++ server/src/routes/auth.route.ts
