+++ server/src/utils/generateToken.ts
